pulled fulfillment record API out into own zome
