+++ Cargo.toml
@@ -7,6 +7,7 @@ members= [
   "lib/vf_planning",
   "lib/vf_knowledge",
   "happs/observation/zomes/economic_event/code",
+  "happs/observation/zomes/fulfillment/code",
   "happs/planning/zomes/commitment/code",
   "happs/planning/zomes/intent/code",
 ]
