+++ happs/observation/zomes/economic_event/code/src/lib.rs
@@ -18,7 +18,6 @@ extern crate serde_json;
 extern crate hdk_graph_helpers;
 extern crate vf_observation;
 mod economic_event_requests;
-mod fulfillment_requests;
 
 use hdk::{
     entry_definition::ValidatingEntryType,
@@ -44,27 +43,13 @@ use vf_observation::economic_event::{
     UpdateRequest as EconomicEventUpdateRequest,
     ResponseData as EconomicEventResponse,
 };
-use vf_planning::fulfillment::{
-    Entry as FulfillmentEntry,
-    CreateRequest as FulfillmentCreateRequest,
-    UpdateRequest as FulfillmentUpdateRequest,
-    ResponseData as FulfillmentResponse,
-};
 use economic_event_requests::{
     handle_get_economic_event,
     handle_create_economic_event,
     handle_update_economic_event,
     handle_delete_economic_event,
 };
-use fulfillment_requests::{
-    handle_create_fulfillment,
-    handle_get_fulfillment,
-    handle_update_fulfillment,
-    handle_delete_fulfillment,
-    handle_query_fulfillments,
-};
 use vf_observation::identifiers::{
-    BRIDGED_PLANNING_DHT,
     EVENT_BASE_ENTRY_TYPE,
     EVENT_ENTRY_TYPE,
     EVENT_FULFILLS_LINK_TYPE,
@@ -72,9 +57,6 @@ use vf_observation::identifiers::{
 };
 use vf_planning::identifiers::{
     FULFILLMENT_BASE_ENTRY_TYPE,
-    FULFILLMENT_ENTRY_TYPE,
-    FULFILLMENT_FULFILLEDBY_LINK_TYPE,
-    FULFILLMENT_FULFILLEDBY_LINK_TAG,
 };
 
 // Zome entry type wrappers
@@ -133,68 +115,12 @@ fn event_base_entry_def() -> ValidatingEntryType {
     )
 }
 
-fn fulfillment_entry_def() -> ValidatingEntryType {
-    entry!(
-        name: FULFILLMENT_ENTRY_TYPE,
-        description: "Represents many-to-many relationships between commitments and economic events that fully or partially satisfy one or more commitments.",
-        sharing: Sharing::Public,
-        validation_package: || {
-            hdk::ValidationPackageDefinition::Entry
-        },
-        validation: |_validation_data: hdk::EntryValidationData<FulfillmentEntry>| {
-            Ok(())
-        }
-    )
-}
-
-fn fulfillment_base_entry_def() -> ValidatingEntryType {
-    entry!(
-        name: FULFILLMENT_BASE_ENTRY_TYPE,
-        description: "Base anchor for initial fulfillment addresses to provide lookup functionality",
-        sharing: Sharing::Public,
-        validation_package: || {
-            hdk::ValidationPackageDefinition::Entry
-        },
-        validation: |_validation_data: hdk::EntryValidationData<Address>| {
-            Ok(())
-        },
-        links: [
-            to!(
-                FULFILLMENT_ENTRY_TYPE,
-                link_type: LINK_TYPE_INITIAL_ENTRY,
-
-                validation_package: || {
-                    hdk::ValidationPackageDefinition::Entry
-                },
-
-                validation: | _validation_data: hdk::LinkValidationData| {
-                    Ok(())
-                }
-            ),
-            to!(
-                EVENT_BASE_ENTRY_TYPE,
-                link_type: FULFILLMENT_FULFILLEDBY_LINK_TYPE,
-
-                validation_package: || {
-                    hdk::ValidationPackageDefinition::Entry
-                },
-
-                validation: | _validation_data: hdk::LinkValidationData| {
-                    Ok(())
-                }
-            )
-        ]
-    )
-}
-
 // Zome definition
 
 define_zome! {
     entries: [
        event_entry_def(),
-       event_base_entry_def(),
-       fulfillment_entry_def(),
-       fulfillment_base_entry_def()
+       event_base_entry_def()
     ]
 
     init: || {
@@ -230,32 +156,6 @@ define_zome! {
             outputs: |result: ZomeApiResult<bool>|,
             handler: handle_delete_economic_event
         }
-
-        create_fulfillment: {
-            inputs: |fulfillment: FulfillmentCreateRequest|,
-            outputs: |result: ZomeApiResult<FulfillmentResponse>|,
-            handler: handle_create_fulfillment
-        }
-        get_fulfillment: {
-            inputs: |address: Address|,
-            outputs: |result: ZomeApiResult<FulfillmentResponse>|,
-            handler: handle_get_fulfillment
-        }
-        update_fulfillment: {
-            inputs: |fulfillment: FulfillmentUpdateRequest|,
-            outputs: |result: ZomeApiResult<FulfillmentResponse>|,
-            handler: handle_update_fulfillment
-        }
-        delete_fulfillment: {
-            inputs: |address: Address|,
-            outputs: |result: ZomeApiResult<bool>|,
-            handler: handle_delete_fulfillment
-        }
-        query_fulfillments: {
-            inputs: |economic_event: Address|,
-            outputs: |result: ZomeApiResult<Vec<FulfillmentResponse>>|,
-            handler: handle_query_fulfillments
-        }
     ]
 
     traits: {
@@ -263,13 +163,7 @@ define_zome! {
             create_event,
             get_event,
             update_event,
-            delete_event,
-
-            create_fulfillment,
-            get_fulfillment,
-            update_fulfillment,
-            delete_fulfillment,
-            query_fulfillments
+            delete_event
         ]
     }
 }
