+++ happs/observation/zomes/economic_event/zome.json
@@ -1,12 +1,3 @@
 {
-  "description": "Recorded observations data for a ValueFlows-compatible economic network",
-  "bridges": [
-    {
-      "presence": "optional",
-      "handle": "vf_planning",
-      "reference": {
-        "traits": {}
-      }
-    }
-  ]
+  "description": "Recorded observations data for a ValueFlows-compatible economic network"
 }
