+++ happs/observation/zomes/fulfillment/code/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "rea_observation_fulfillment"
+version = "0.1.0"
+authors = ["pospi <pospi@spadgos.com>"]
+edition = "2018"
+
+[dependencies]
+serde = "=1.0.89"
+serde_json = { version = "=1.0.39", features = ["preserve_order"] }
+serde_derive = "=1.0.89"
+# :DUPE: hdk-rust-revid
+hdk = { git = "https://github.com/holochain/holochain-rust", tag = "0.0.28-alpha1" }
+holochain_wasm_utils = { git = "https://github.com/holochain/holochain-rust", tag = "0.0.28-alpha1" }
+holochain_json_derive = { version = "0.0.1-alpha2" }
+
+hdk_graph_helpers = { path = "../../../../../lib/hdk_graph_helpers" }
+vf_observation = { path = "../../../../../lib/vf_observation" }
+vf_planning = { path = "../../../../../lib/vf_planning" }
+
+[lib]
+path = "src/lib.rs"
+crate-type = ["cdylib"]
